+++ ChangeLog.rst
@@ -6,6 +6,7 @@ UNRELEASED
 * Add support for Python 3.9.
 * Fix extra line breaks inside html link text (between '[' and ']')
 * Fix #344: indent ``<ul>`` inside ``<ol>`` three spaces instead of two to comply with CommonMark, GFM, etc.
+* Don't wrap tables by default and add a ``--wrap-tables`` config option
 
 2020.1.16
 =========